requirements.txt
amqp==2.5.0
authy==2.2.3
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-extensions==2.1.7
django-grappelli==2.12.3
django-leaflet==0.24.0
django-modeltranslation==0.13.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
kombu==4.6.1
Pillow==6.0.0
psycopg2-binary==2.8.2
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
